feat:added flask and established connection
unfinished yung home since nguba
