+++ poster_fetcher.py
@@ -1,0 +1,18 @@
+import requests
+
+TMDB_API_KEY = "a9d902d7338397e2deda22858c168821"
+
+def get_poster_url(movie_title):
+    search_url = "https://api.themoviedb.org/3/search/movie"
+    params = {
+        'api_key': TMDB_API_KEY,
+        'query': movie_title
+    }
+    response = requests.get(search_url, params=params)
+    if response.status_code == 200:
+        results = response.json().get("results")
+        if results:
+            poster_path = results[0].get("poster_path")
+            if poster_path:
+                return f"https://image.tmdb.org/t/p/w500{poster_path}"
+    return None
