+++ requirements.txt
@@ -14,3 +14,4 @@ sklearn==0.0
 threadpoolctl==3.6.0
 tzdata==2025.2
 urllib3==2.4.0
+flask
