+++ templates/navbar.html
@@ -1,0 +1,18 @@
+ <!-- Nav -->
+    <nav class="text-center p-6 flex items-center justify-between z-3 px-[8%]">
+        <div>
+            <h1 class="font-title text-[2rem]">Rekom</h1>
+        </div>
+        <div class="flex justify-center flex-1 gap-20 font-body c1 text-lg">
+            <a href="/" class="text-[#FD5017] transition hover:text-[#4EC3B9]">Home</a>
+            <a href="/rekom" class="transition hover:text-[#4EC3B9]">Reko</a>
+            <a href="/about" class="transition hover:text-[#4EC3B9]">About Us</a>
+            
+
+        </div>
+        <div>
+            <a href="/contact" class="transition flex"> <img src="../static/images/headset.svg" alt="headset"
+       class="w-8 h-8 invert transition hover:text-[#4EC3B9]" />
+</a>
+        </div>
+    </nav>
