+++ templates/rekom.html
@@ -1,0 +1,30 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Rekom</title>
+    <script src="https://cdn.jsdelivr.net/npm/@tailwindcss/browser@4"></script>
+    <link rel="stylesheet" href="../static/css/styles.css" />
+</head>
+<body>
+  <h1>🎬 Movie Recommendation System</h1>
+
+  <label for="genres">Select Genres:</label>
+  <select id="genres" multiple>
+    <!-- Options will be added dynamically -->
+  </select>
+
+  <label for="user_id">User ID (optional):</label>
+  <input type="text" id="user_id" placeholder="1 to 6040">
+
+  <label for="title">Movie Title (optional):</label>
+  <input type="text" id="title">
+
+  <button onclick="getRecommendations()">Get Recommendations</button>
+
+  <div id="results"></div>
+
+  <script src="/static/app.js"></script>
+</body>
+</html>
